kpme-2111: use Balance transfer document to transfer SSTO usage from current leave calendar
git-svn-id: https://svn.kuali.org/repos/kpme/trunk@4126 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ src/main/java/org/kuali/hr/lm/leavecalendar/web/LeaveCalendarAction.java
@@ -437,6 +437,13 @@ public class LeaveCalendarAction extends TkAction {
         		}
         	}
         	
+        	if(TkServiceLocator.getPermissionsService().canBankOrTransferSSTOUsage(blockToDelete)) {
+        		// prompt the user to do a create a balance transfer document
+        		ActionRedirect redirect = new ActionRedirect();
+        		redirect.setPath("/BalanceTransfer.do?"+ request.getQueryString());
+        		return redirect;
+        	}
+        	
         	TkServiceLocator.getLeaveBlockService().deleteLeaveBlock(leaveBlockId, principalId);
 		    generateLeaveCalendarChangedNotification(principalId, targetPrincipalId, documentId, calendarEntry.getHrCalendarEntriesId());
 		    if(CollectionUtils.isNotEmpty(approverList)) {
